Use experimental::filesystem for the power cap
Using std::filesystem changes the behavior of the path::/ operator
compared to std::experimental::filesystem, and this doesn't work
with the existing code. Specifically, appending a path that starts with
a '/' which is done here completely overwrites the base path instead
of appending.

Change back to using
std::experimental::filesytem to keep things consistent with the rest of
the code base.  If a move to std::filesystem is desired in the future the
whole codebase can be changed and tested at once.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I2b1cbbbc0ab13f1a0a5bf85494302e68e7adcab3